chore(xlsx): extract chart plot-area host into chart/plotArea
Move every parseAxes / parsePlotArea* / parseVaryColors /
parseScatterStyle / parseDropLines / parseHiLowLines / parseSerLines /
parseShowLineMarkers / parseBarGrouping / parseHoleSize / parseGapWidth /
parseUpDownBarsGapWidth / parseOverlap / parseFirstSliceAng /
parseLineAreaGrouping helper from chart-reader.ts into
chart/plotArea.ts. Move every buildPlotArea / buildPlotAreaSpPr /
buildBarChart / buildLineChart / buildAreaChart / buildPieChart /
buildDoughnutChart / buildScatterChart / buildUpDownBars /
clamp{GapWidth,Overlap,UpDownBarsGapWidth,FirstSliceAng,HoleSize} /
normalizePlotAreaFillColor / normalizePlotAreaBorderColor /
resolvePlotAreaLayout / resolveVaryColors / resolveScatterStyle helper
from chart-writer.ts into chart/plotArea.ts.

Move the writer-side DOUGHNUT_HOLE_DEFAULT/MIN/MAX,
VARY_COLORS_DEFAULT_TRUE_TYPES, AXIS_TITLE_DEFAULT_FONT_SIZE_SZ
duplicate, SCATTER_STYLE_VALUES, and resolvePlotAreaLayout out of
chart-writer.ts since they were only used by the moved plot-area
functions. Move the reader-side VARY_COLORS_DEFAULT_TRUE and
VALID_SCATTER_STYLES out of chart-reader.ts for the same reason.

The plot-area module imports buildBarAxes / buildScatterAxes /
AXIS_ID_* / AxisRenderOptions from chart/axis, buildDataTable /
resolveDataTable from chart/dataTable, buildChartLevelDataLabels from
chart/dataLabels, buildSeries from chart/series — composing the
plot-area subtree from its host children.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
