+++ src/xlsx/chart-reader.ts
@@ -103,6 +103,26 @@ import { parseDataTable } from "./chart/dataTable";
 import { parseDataLabels } from "./chart/dataLabels";
 import { parseSeries } from "./chart/series";
 import { parseAutoTitleDeleted, parseAxisInfo } from "./chart/axis";
+import {
+  parseAxes,
+  parseBarGrouping,
+  parseDropLines,
+  parseFirstSliceAng,
+  parseGapWidth,
+  parseHiLowLines,
+  parseHoleSize,
+  parseLineAreaGrouping,
+  parseOverlap,
+  parsePlotAreaBorderColor,
+  parsePlotAreaBorderWidth,
+  parsePlotAreaFillColor,
+  parsePlotAreaLayout,
+  parseScatterStyle,
+  parseSerLines,
+  parseShowLineMarkers,
+  parseUpDownBarsGapWidth,
+  parseVaryColors,
+} from "./chart/plotArea";
 
 /** All chart-type element local names recognized by Excel. */
 const CHART_KIND_TAGS: ReadonlyMap<string, ChartKind> = new Map([
@@ -862,61 +882,6 @@ export function parseChart(xml: string): Chart | undefined {
 
 // ── Axes ──────────────────────────────────────────────────────────
 
-/**
- * Pull per-axis metadata from the plot area's `<c:catAx>` / `<c:valAx>`
- * children.
- *
- * The mapping mirrors the writer side:
- *   - bar / column / line / area: `x` = `<c:catAx>`, `y` = first `<c:valAx>`.
- *   - scatter / bubble:           `x` = first `<c:valAx>`, `y` = second `<c:valAx>`.
- *
- * Returns `undefined` when neither axis surfaces a title — keeps the
- * default `Chart` shape lean.
- */
-function parseAxes(plotArea: XmlElement): { x?: ChartAxisInfo; y?: ChartAxisInfo } | undefined {
-  let catAx: XmlElement | undefined;
-  const valAxes: XmlElement[] = [];
-  for (const child of childElements(plotArea)) {
-    if (child.local === "catAx") {
-      catAx ??= child;
-    } else if (child.local === "valAx") {
-      valAxes.push(child);
-    }
-  }
-
-  let xAxis: XmlElement | undefined;
-  let yAxis: XmlElement | undefined;
-  if (catAx) {
-    xAxis = catAx;
-    yAxis = valAxes[0];
-  } else {
-    // Scatter / bubble: both axes are valAx. The first declared one is
-    // the X axis (`axPos="b"`), the second is the Y axis (`axPos="l"`).
-    xAxis = valAxes[0];
-    yAxis = valAxes[1];
-  }
-
-  // `<c:crossBetween>` is required on every `<c:valAx>` — Excel always
-  // emits the element — so the reader needs a per-family default to
-  // collapse against. The catAx-anchored families (bar / column / line
-  // / area) emit `"between"` on the value axis; scatter (catAx-less,
-  // both axes are valAx) emits `"midCat"` on both axes. Pass the
-  // expected default to `parseAxisInfo` so a chart that inherited the
-  // default round-trips identically through {@link cloneChart} —
-  // absence on the parsed shape and the writer-emitted default produce
-  // the same `<c:crossBetween val=".."/>` byte-for-byte.
-  const familyDefaultCrossBetween: ChartAxisCrossBetween = catAx ? "between" : "midCat";
-
-  const x = xAxis ? parseAxisInfo(xAxis, familyDefaultCrossBetween) : undefined;
-  const y = yAxis ? parseAxisInfo(yAxis, familyDefaultCrossBetween) : undefined;
-
-  if (!x && !y) return undefined;
-  const out: { x?: ChartAxisInfo; y?: ChartAxisInfo } = {};
-  if (x) out.x = x;
-  if (y) out.y = y;
-  return out;
-}
-
 
 /**
  * Pull a finite numeric `val=".."` attribute off a named child of
@@ -1019,134 +984,6 @@ function formulaText(wrapper: XmlElement | undefined): string | undefined {
 
 // ── Legend ────────────────────────────────────────────────────────
 
-
-/**
- * Pull `<c:plotArea><c:layout><c:manualLayout>` off the chart. Reflects
- * Excel's "Format Plot Area -> Position -> Custom" placement — the
- * `(x, y)` anchor and `(w, h)` size of the plot area as fractions of
- * the chart frame in the `0..1` band.
- *
- * The OOXML schema (`CT_PlotArea`, ECMA-376 Part 1, §21.2.2.145) places
- * `<c:layout>` as the first child of `<c:plotArea>`, before any
- * chart-type element / axes / `<c:dTable>` / `<c:spPr>`. The
- * `<c:manualLayout>` block (`CT_ManualLayout`, §21.2.2.115) exposes
- * optional `<c:x>` / `<c:y>` / `<c:w>` / `<c:h>` children whose `val`
- * attributes carry an `xsd:double`. The reader admits the coordinate
- * only when `val` parses to a finite number in the `0..1` band;
- * out-of-range / non-finite / non-numeric tokens drop to `undefined`
- * on the matching axis so absence and a malformed token round-trip
- * identically through {@link cloneChart}.
- *
- * Both `<c:xMode val="edge"/>` (absolute fraction of the chart frame)
- * and `<c:xMode val="factor"/>` (delta from auto-layout) are accepted —
- * the reader surfaces the same `ChartManualLayout` shape regardless,
- * since the writer always normalizes to `"edge"` on emit (Excel itself
- * emits the absolute form when the user drags an element to a custom
- * position).
- *
- * Returns `undefined` whenever the plot area omits the `<c:layout>` /
- * `<c:manualLayout>` chain at any link (the bare `<c:layout/>`
- * placeholder Excel emits on auto-layout charts has no `<c:manualLayout>`
- * child, so it collapses cleanly here), or when every coordinate
- * dropped on normalization — the field is omitted entirely on a clean
- * parse so absence and an empty layout round-trip identically through
- * the writer.
- */
-function parsePlotAreaLayout(plotArea: XmlElement): ChartManualLayout | undefined {
-  return parseManualLayout(plotArea);
-}
-
-/**
- * Pull `<c:plotArea><c:spPr><a:solidFill><a:srgbClr val=".."/></a:solidFill>
- * </c:spPr></c:plotArea>` off the plot area. Returns the plot-area solid
- * fill color as a 6-character uppercase hex string.
- *
- * The OOXML `<a:srgbClr>` element carries the literal sRGB color
- * (`CT_SRgbColor`, ECMA-376 Part 1, §20.1.2.3.32) inside the fill choice
- * of `<c:spPr>` (`CT_ShapeProperties`, §20.1.2.3.13). The `<c:spPr>` slot
- * sits at the tail of `<c:plotArea>` per `CT_PlotArea` (§21.2.2.145),
- * after every chart-type element / axes / `<c:dTable>`.
- *
- * The reader surfaces only the literal `<a:srgbClr>` form — absence,
- * non-solid fills (`<a:noFill>` / `<a:gradFill>` / `<a:pattFill>` /
- * `<a:blipFill>`), and theme-color references (`<a:schemeClr>`) all
- * collapse to `undefined` so a chart that pinned a fill the writer
- * cannot reproduce on emit drops the field rather than fabricate one
- * Excel would render differently. Malformed `val` tokens (wrong length,
- * non-hex characters, alpha-channel forms, non-string escapes) likewise
- * drop to `undefined`.
- *
- * Mirrors the writer-side {@link SheetChart.plotAreaFillColor} so a
- * parsed value slots straight into {@link cloneChart} without
- * conversion. The lookup is scoped to direct children of `<c:plotArea>`
- * so a stray `<c:spPr>` elsewhere (e.g. on a series, on an axis, on the
- * `<c:dTable>` block) cannot leak in.
- */
-function parsePlotAreaFillColor(plotArea: XmlElement): string | undefined {
-  return parseSpPrFill(plotArea);
-}
-
-/**
- * Pull `<c:plotArea><c:spPr><a:ln><a:solidFill><a:srgbClr val=".."/>
- * </a:solidFill></a:ln></c:spPr></c:plotArea>` off the plot-area block.
- * Returns the line stroke color as a 6-character uppercase hex string.
- *
- * The OOXML `<a:srgbClr>` element carries the literal sRGB color
- * (`CT_SRgbColor`, ECMA-376 Part 1, §20.1.2.3.32) inside the line's
- * solid fill choice (`CT_LineProperties`, §20.1.2.3.24) which itself
- * sits inside `<c:spPr>` (`CT_ShapeProperties`, §20.1.2.3.13). The
- * `<c:spPr>` slot lives at the tail of `<c:plotArea>` per
- * `CT_PlotArea` (§21.2.2.145), after every chart-type element / axes /
- * `<c:dTable>`.
- *
- * The reader surfaces only the literal `<a:srgbClr>` form — absence,
- * non-solid line fills (`<a:noFill>` / `<a:gradFill>` / `<a:pattFill>`),
- * and theme-color references (`<a:schemeClr>`) all collapse to
- * `undefined` so a chart that pinned a stroke the writer cannot
- * reproduce on emit drops the field rather than fabricate one Excel
- * would render differently. Malformed `val` tokens (wrong length,
- * non-hex characters, alpha-channel forms, non-string escapes)
- * likewise drop to `undefined`.
- *
- * Mirrors the writer-side {@link SheetChart.plotAreaBorderColor} so a
- * parsed value slots straight into {@link cloneChart} without
- * conversion. The lookup is scoped to direct children of
- * `<c:plotArea>` so a stray `<c:spPr>` elsewhere (e.g. on a series,
- * on an axis, on the `<c:dTable>` block) cannot leak in. Mirrors
- * {@link parsePlotAreaFillColor} — same `<c:spPr>` host element on
- * the same `<c:plotArea>` parent — but lands on the line
- * (`<a:ln><a:solidFill>`) child rather than the fill
- * (`<a:solidFill>`) child.
- */
-function parsePlotAreaBorderColor(plotArea: XmlElement): string | undefined {
-  return parseSpPrBorderColor(plotArea);
-}
-
-/**
- * Pull the `w` attribute off `<c:plotArea><c:spPr><a:ln w="EMU"/>` and
- * return the stroke width in points after clamping to the
- * `0.25..13.5` pt band Excel's UI exposes. The OOXML `w` attribute
- * carries the stroke width in English Metric Units (1 pt = 12 700 EMU)
- * per `CT_LineProperties` (ECMA-376 Part 1, §20.1.2.3.24); the reader
- * snaps the result to the 0.25 pt grid so a parsed-then-written width
- * does not drift across round-trips (Excel rounds in the UI anyway).
- *
- * Returns `undefined` when there is no `<c:spPr><a:ln w=..>` slot, when
- * the attribute is missing, when the value cannot be parsed as a finite
- * positive number, or when it parses to zero (Excel's "no border"
- * marker — the writer-side knob does not model that state). Mirrors
- * the writer-side {@link SheetChart.plotAreaBorderWidth} so a parsed
- * value slots straight into {@link cloneChart} without conversion.
- *
- * The lookup is scoped to direct children of `<c:plotArea>` so a stray
- * `<a:ln w=..>` elsewhere (on a series stroke, on an axis line) cannot
- * leak in. Mirrors {@link parsePlotAreaBorderColor} — same `<c:spPr>`
- * host on the same `<c:plotArea>` parent — but lands on the `w`
- * attribute rather than the `<a:solidFill><a:srgbClr>` color child.
- */
-function parsePlotAreaBorderWidth(plotArea: XmlElement): number | undefined {
-  return parseBorderWidthFromSpPr(plotArea);
-}
 
 /**
  * Pull `<c:chartSpace><c:spPr><a:solidFill><a:srgbClr val=".."/>
@@ -1605,368 +1442,17 @@ function parseProtectionFlag(protection: XmlElement, local: string): boolean | u
 
 // ── Vary Colors ────────────────────────────────────────────────────
 
-/**
- * Chart kinds that default `<c:varyColors>` to `1` in OOXML — every
- * data point in the (single) series carries a unique color. Excel's
- * pie / doughnut / ofPie templates emit `<c:varyColors val="1"/>` so
- * absence and `1` collapse to `undefined` here; only an explicit `0`
- * surfaces `false`.
- */
-const VARY_COLORS_DEFAULT_TRUE: ReadonlySet<ChartKind> = new Set([
-  "pie",
-  "pie3D",
-  "doughnut",
-  "ofPie",
-]);
-
-/**
- * Pull `<c:varyColors val=".."/>` off a chart-type element.
- *
- * Excel's per-family default flips the meaning: pie / doughnut /
- * pie3D / ofPie default to `true` (every slice unique) while every
- * other chart family defaults to `false` (one color per series).
- * Matching values collapse to `undefined` so a roundtrip of a stock
- * template stays minimal — only non-default values surface so
- * {@link cloneChart} can carry them through. Unknown values and
- * missing `val` attributes drop to `undefined`.
- */
-function parseVaryColors(chartTypeEl: XmlElement, kind: ChartKind): boolean | undefined {
-  const el = findChild(chartTypeEl, "varyColors");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  const familyDefaultsTrue = VARY_COLORS_DEFAULT_TRUE.has(kind);
-  // Accept the OOXML truthy / falsy spellings. `1` / `true` map to true,
-  // `0` / `false` map to false, anything else drops.
-  let parsed: boolean;
-  switch (raw) {
-    case "1":
-    case "true":
-      parsed = true;
-      break;
-    case "0":
-    case "false":
-      parsed = false;
-      break;
-    default:
-      return undefined;
-  }
-  // Collapse the per-family default so absence and the default
-  // round-trip identically.
-  if (parsed === familyDefaultsTrue) return undefined;
-  return parsed;
-}
-
-// ── Scatter Style ─────────────────────────────────────────────────
-
-/**
- * Recognized values of `<c:scatterStyle>` per the OOXML
- * `ST_ScatterStyle` enumeration. Tokens outside the set drop to
- * `undefined` so a corrupt template does not surface a string Excel
- * would not emit.
- */
-const VALID_SCATTER_STYLES: ReadonlySet<ChartScatterStyle> = new Set([
-  "none",
-  "line",
-  "lineMarker",
-  "marker",
-  "smooth",
-  "smoothMarker",
-]);
-
-/**
- * Pull `<c:scatterStyle val=".."/>` off a `<c:scatterChart>` element.
- *
- * The OOXML schema lists the element as required on `<c:scatterChart>`
- * but tolerates absence in practice — Excel falls back to `"marker"`
- * (the schema default per CT_ScatterStyle) when the file omits it.
- * The reader does not pin a default of its own: every literal value
- * in {@link VALID_SCATTER_STYLES} surfaces as-is so a clone preserves
- * the exact preset the template authored. Missing elements, missing
- * `val` attributes, and tokens outside the enum drop to `undefined`.
- *
- * Note that the writer's default is `"lineMarker"` (Excel's chart-
- * picker default and what every fresh hucre scatter chart emits today),
- * which differs from the OOXML schema default of `"marker"`. The
- * asymmetry is intentional — writing `"lineMarker"` matches Excel's
- * UI default; not collapsing it on read keeps the round-trip exact.
- */
-function parseScatterStyle(scatterChart: XmlElement): ChartScatterStyle | undefined {
-  const el = findChild(scatterChart, "scatterStyle");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  if (!VALID_SCATTER_STYLES.has(raw as ChartScatterStyle)) return undefined;
-  return raw as ChartScatterStyle;
-}
-
-// ── Drop Lines / Hi-Low Lines ─────────────────────────────────────
-
-/**
- * Pull `<c:dropLines/>` off a `<c:lineChart>` / `<c:line3DChart>` /
- * `<c:areaChart>` / `<c:area3DChart>` element. Returns `true` when
- * the element is present (its mere presence paints the connector
- * lines per OOXML CT_ChartLines), `undefined` otherwise so absence
- * collapses to the writer's default.
- *
- * `<c:dropLines>` is structurally a `CT_ChartLines` and may carry a
- * nested `<c:spPr>` for stroke styling, but hucre's reader only
- * surfaces the on/off bit — the shape properties are not modelled in
- * this phase. A template that pins custom drop-line colors / widths
- * therefore round-trips as a default-styled line; the on/off intent
- * still survives, which is what {@link cloneChart} needs.
- */
-function parseDropLines(chartTypeEl: XmlElement): boolean | undefined {
-  return findChild(chartTypeEl, "dropLines") ? true : undefined;
-}
-
-/**
- * Pull `<c:hiLowLines/>` off a `<c:lineChart>` / `<c:line3DChart>` /
- * `<c:stockChart>` element. Same on/off shape as
- * {@link parseDropLines}; the element is bare so its mere presence
- * surfaces `true`, absence collapses to `undefined`.
- */
-function parseHiLowLines(chartTypeEl: XmlElement): boolean | undefined {
-  return findChild(chartTypeEl, "hiLowLines") ? true : undefined;
-}
-
-/**
- * Pull `<c:serLines/>` off a `<c:barChart>` / `<c:ofPieChart>` element.
- * Same on/off shape as {@link parseDropLines} / {@link parseHiLowLines};
- * the element is bare so its mere presence surfaces `true`, absence
- * collapses to `undefined`.
- *
- * `<c:serLines>` is structurally a `CT_ChartLines` and may carry a
- * nested `<c:spPr>` for stroke styling, but hucre's reader only
- * surfaces the on/off bit at this layer (mirrors how `parseDropLines`
- * handles the same shape on its hosts). Even when the nested `<c:spPr>`
- * is the only child, the presence flag still survives, which is what
- * {@link cloneChart} needs.
- */
-function parseSerLines(chartTypeEl: XmlElement): boolean | undefined {
-  return findChild(chartTypeEl, "serLines") ? true : undefined;
-}
-
-// ── Chart-level Marker Visibility ─────────────────────────────────
-
-/**
- * Pull `<c:marker val=".."/>` off a `<c:lineChart>` element. This is
- * the chart-level CT_Boolean variant of `<c:marker>` — distinct from
- * the per-series `<c:marker>` (CT_Marker, with style / size / fill).
- * The element gates whether per-series markers paint at all on the
- * line chart.
- *
- * The OOXML / Excel default `val="1"` (markers shown) collapses to
- * `undefined` so absence and the default round-trip identically
- * through {@link cloneChart}; only an explicit `val="0"` surfaces
- * `false`. Accepts the OOXML truthy / falsy spellings (`"1"` /
- * `"true"` / `"0"` / `"false"`); unknown values, missing `val`
- * attributes, and a missing element all drop to `undefined`.
- *
- * The chart-level slot lives exclusively on `CT_LineChart` per the
- * OOXML schema — `CT_Line3DChart` and `CT_StockChart` have no
- * chart-level marker toggle. Caller is expected to gate the lookup
- * on the matching chart-type kind.
- */
-function parseShowLineMarkers(lineChart: XmlElement): boolean | undefined {
-  const el = findChild(lineChart, "marker");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  switch (raw) {
-    case "0":
-    case "false":
-      return false;
-    case "1":
-    case "true":
-      // OOXML / Excel default — collapse to undefined for symmetry
-      // with the writer's `showLineMarkers` field, so a fresh chart
-      // and a marker-on chart round-trip identically.
-      return undefined;
-    default:
-      return undefined;
-  }
-}
-
 // ── Bar Grouping ──────────────────────────────────────────────────
 
-/**
- * Pull `<c:grouping val=".."/>` off a `<c:barChart>` element. Returns
- * `undefined` when the grouping element is missing or carries the
- * default `"standard"` / `"clustered"` value — the writer's
- * {@link SheetChart.barGrouping} treats both as the unspecified
- * default, so omitting them keeps the parsed shape minimal.
- */
-function parseBarGrouping(barChart: XmlElement): ChartBarGrouping | undefined {
-  const grouping = findChild(barChart, "grouping");
-  if (!grouping) return undefined;
-  const val = grouping.attrs.val;
-  if (typeof val !== "string") return undefined;
-  switch (val) {
-    case "stacked":
-      return "stacked";
-    case "percentStacked":
-      return "percentStacked";
-    case "clustered":
-      return "clustered";
-    case "standard":
-      // OOXML's `standard` for barChart is functionally equivalent to
-      // `clustered` (Excel renders side-by-side). Surface neither so
-      // the cloned chart inherits the writer's default.
-      return undefined;
-    default:
-      return undefined;
-  }
-}
 
 // ── Doughnut Hole ─────────────────────────────────────────────────
 
-/**
- * Pull `<c:holeSize val=".."/>` off a `<c:doughnutChart>` element.
- * Returns `undefined` when the attribute is missing, malformed, or
- * outside the 1–99 range OOXML allows. Excel itself only writes values
- * in 10–90 (the UI clamps to that band) but the spec is wider, so we
- * accept the full schema range and let the writer re-clamp on the way
- * back out.
- */
-function parseHoleSize(doughnut: XmlElement): number | undefined {
-  const el = findChild(doughnut, "holeSize");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  const parsed = Number.parseInt(raw, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  if (parsed < 1 || parsed > 99) return undefined;
-  return parsed;
-}
 
 // ── Bar / Column gap width & overlap ──────────────────────────────
 
-/**
- * Pull `<c:gapWidth val=".."/>` off a `<c:barChart>` / `<c:bar3DChart>`
- * element.
- *
- * The OOXML schema (`ST_GapAmount`) restricts the value to the
- * inclusive `0..500` band; out-of-range values are dropped rather than
- * clamped so a corrupt template does not silently rewrite as a
- * different gap. The OOXML default of `150` collapses to `undefined`
- * for symmetry with the writer's {@link SheetChart.gapWidth} default
- * — absence and `150` mean the same thing.
- */
-function parseGapWidth(barChart: XmlElement): number | undefined {
-  const el = findChild(barChart, "gapWidth");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  const parsed = Number.parseInt(raw, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  if (parsed < 0 || parsed > 500) return undefined;
-  if (parsed === 150) return undefined;
-  return parsed;
-}
-
-/**
- * Pull `<c:gapWidth val=".."/>` off a `<c:upDownBars>` element. The
- * value controls the spacing between the up / down bars themselves on
- * a line chart (distinct from the bar-chart `<c:gapWidth>` which
- * controls spacing between category groups).
- *
- * The OOXML schema (`ST_GapAmount`) restricts the value to the
- * inclusive `0..500` band; out-of-range values are dropped rather than
- * clamped so a corrupt template does not silently rewrite as a
- * different gap. The OOXML default of `150` collapses to `undefined`
- * for symmetry with the writer's {@link SheetChart.upDownBarsGapWidth}
- * default — absence and `150` mean the same thing.
- */
-function parseUpDownBarsGapWidth(upDownBars: XmlElement): number | undefined {
-  const el = findChild(upDownBars, "gapWidth");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  const parsed = Number.parseInt(raw, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  if (parsed < 0 || parsed > 500) return undefined;
-  if (parsed === 150) return undefined;
-  return parsed;
-}
-
-/**
- * Pull `<c:overlap val=".."/>` off a `<c:barChart>` / `<c:bar3DChart>`
- * element.
- *
- * The OOXML schema (`ST_Overlap`) restricts the value to the inclusive
- * `-100..100` band; out-of-range values are dropped rather than
- * clamped. The OOXML default of `0` collapses to `undefined` for
- * symmetry with the writer's {@link SheetChart.overlap} default. Note
- * that Excel's reference serialization emits `<c:overlap val="100"/>`
- * for stacked charts even though the schema default is `0`; we surface
- * the literal value carried by the file rather than try to invert
- * Excel's per-grouping default — `100` on a stacked chart therefore
- * round-trips as `100`.
- */
-function parseOverlap(barChart: XmlElement): number | undefined {
-  const el = findChild(barChart, "overlap");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  const parsed = Number.parseInt(raw, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  if (parsed < -100 || parsed > 100) return undefined;
-  if (parsed === 0) return undefined;
-  return parsed;
-}
 
 // ── First Slice Angle ─────────────────────────────────────────────
 
-/**
- * Pull `<c:firstSliceAng val=".."/>` off a `<c:pieChart>` /
- * `<c:doughnutChart>` element. Returns `undefined` when the attribute
- * is missing, malformed, or carries the OOXML default of `0` — the
- * writer's {@link SheetChart.firstSliceAng} treats absence and `0`
- * identically, so collapsing here keeps the round-trip stable.
- *
- * The OOXML schema (CT_FirstSliceAng) restricts the value to the
- * inclusive range `0..360`; out-of-range values are dropped rather
- * than clamped so a corrupt template does not silently rewrite as a
- * different angle.
- */
-function parseFirstSliceAng(chartType: XmlElement): number | undefined {
-  const el = findChild(chartType, "firstSliceAng");
-  if (!el) return undefined;
-  const raw = el.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  const parsed = Number.parseInt(raw, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  if (parsed < 0 || parsed > 360) return undefined;
-  // Collapse `0` and the schema-equivalent `360` to undefined — both
-  // mean "first slice at 12 o'clock", which is the writer's default.
-  if (parsed === 0 || parsed === 360) return undefined;
-  return parsed;
-}
-
-/**
- * Pull `<c:grouping val=".."/>` off a `<c:lineChart>` or `<c:areaChart>`
- * element. Returns `undefined` when the grouping element is missing or
- * carries the default `"standard"` value — the writer's
- * {@link SheetChart.lineGrouping} / {@link SheetChart.areaGrouping}
- * treat that as the absence of the field.
- */
-function parseLineAreaGrouping(chartType: XmlElement): ChartLineAreaGrouping | undefined {
-  const grouping = findChild(chartType, "grouping");
-  if (!grouping) return undefined;
-  const val = grouping.attrs.val;
-  if (typeof val !== "string") return undefined;
-  switch (val) {
-    case "stacked":
-      return "stacked";
-    case "percentStacked":
-      return "percentStacked";
-    case "standard":
-      return undefined;
-    default:
-      return undefined;
-  }
-}
 
 /**
  * Parse an OOXML boolean attribute. The spec allows `"1"` / `"0"` /
